Added Dapper library and related changes
In this commit, we've made several changes related to the addition of the Dapper library. In `RowPositionWorker.cs`, we've added Dapper to the list of using directives and introduced a new method `RigSuppliersTable` in the `SqlUtilities` namespace for setting up a SQL database connection. We've also added a reference to Dapper in `SqlUtilities.csproj` and included a `packages.config` file to list the project's NuGet dependencies, which now includes Dapper version `2.1.35`. A new file `DapperOperations.cs` was added, containing a `DapperOperations` class with a `CategoryIdentifiersAsync` method for retrieving category identifiers from a SQL database. Additionally, we've made some changes in `Form1.cs` and `Form1.resx` files, including a syntax correction and removal of metadata entries related to `UserAddedColumn`.